plan: distinguish the physical plan test and logical plan test.
